s4:dcerpc - fix DEBUG output warning

+++ source4/librpc/rpc/dcerpc.c
@@ -457,9 +457,10 @@ static NTSTATUS ncacn_push_request_sign(struct dcerpc_connection *c,
 
 	if (creds2.length != sig_size) {
 		DEBUG(0,("ncacn_push_request_sign: creds2.length[%u] != sig_size[%u] pad[%u] stub[%u]\n",
-			creds2.length, (uint32_t)sig_size,
-			c->security_state.auth_info->auth_pad_length,
-			pkt->u.request.stub_and_verifier.length));
+			(unsigned) creds2.length,
+			(unsigned) sig_size,
+			(unsigned) c->security_state.auth_info->auth_pad_length,
+			(unsigned) pkt->u.request.stub_and_verifier.length));
 		return NT_STATUS_INTERNAL_ERROR;
 	}
